result/server/pyproject.toml
[tool.poetry]
name = "blockscrape-result-server"
version = "0.1.0"
description = ""
authors = ["Lars Jaeger <lars@jgr.one>"]
readme = "README.md"
packages = [{include = "blockscrape_result_server"}]

[tool.poetry.dependencies]
python = "^3.10"
redis = {extras = ["hiredis"], version = "^4.6.0"}
python-socketio = "^5.9.0"
pydantic = "^2.3.0"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
pika = "^1.3.2"
eventlet = "^0.33.3"

[tool.poetry.group.dev.dependencies]
setuptools = "^68.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
